feat: implement drag-and-drop functionality for widget reordering in content section
